fix(local-mode): classify plaintext LAN refresh imports as access-only
The refresh token is only ever sent over https or loopback, so a bundle
whose gateway is non-loopback plaintext http can never renew. Report it
access-only so the expiry warning shows. The confidential-channel
predicate moves into the package; the CLI imports it.
